extract `GenericTypeParameterMapper` to its own file
