+++ src/Tapper/DefaultTypeMapperProvider.cs
@@ -6,19 +6,6 @@ using Tapper.TypeMappers;
 
 namespace Tapper;
 
-internal class GenericTypeParameterMapper : ITypeMapper
-{
-    public ITypeSymbol Assign { get; } = default!;
-
-    public string MapTo(ITypeSymbol typeSymbol, ITranspilationOptions options)
-    {
-        if (typeSymbol is not ITypeParameterSymbol typeParameterSymbol)
-            throw new InvalidOperationException($"GenericTypeMapper does not support {typeSymbol.ToDisplayString()}.");
-
-        return typeParameterSymbol.Name;
-    }
-}
-
 public class DefaultTypeMapperProvider : ITypeMapperProvider
 {
     private readonly ArrayTypeMapper _arrayTypeMapper;
